+++ app/app.component.spec.ts
@@ -1,14 +1,9 @@
 /* tslint:disable:no-unused-variable */
 import { AppComponent } from './app.component';
 
-import { async, inject } from '@angular/core/testing';
-
-import { TestComponentBuilder } from '@angular/core/testing';
+import { TestBed } from '@angular/core/testing';
 
 import { By }             from '@angular/platform-browser';
-import { provide }        from '@angular/core';
-import { ViewMetadata }   from '@angular/core';
-import { PromiseWrapper } from '@angular/core/src/facade/promise';
 
 ////////  SPECS  /////////////
 
@@ -20,27 +15,23 @@ describe('Smoke test', () => {
 });
 
 describe('AppComponent with TCB', function () {
+  beforeEach(() => {
+    TestBed.configureTestingModule({declarations: [AppComponent]});
+  });
 
-  it('should instantiate component',
-    async(inject([TestComponentBuilder], (tcb: TestComponentBuilder) => {
+  it('should instantiate component', () => {
+    let fixture = TestBed.createComponent(AppComponent);
+    expect(fixture.componentInstance instanceof AppComponent).toBe(true, 'should create AppComponent');
+  });
 
-    tcb.createAsync(AppComponent).then(fixture => {
-      expect(fixture.componentInstance instanceof AppComponent).toBe(true, 'should create AppComponent');
-    });
-  })));
+  it('should have expected <h1> text', () => {
+    let fixture = TestBed.createComponent(AppComponent);
+    fixture.detectChanges();
 
-  it('should have expected <h1> text',
-    async(inject([TestComponentBuilder], (tcb: TestComponentBuilder) => {
+    let h1 = fixture.debugElement.query(el => el.name === 'h1').nativeElement;  // it works
 
-      tcb.createAsync(AppComponent).then(fixture => {
-      // fixture.detectChanges();  // would need to resolve a binding but we don't have a binding
+        h1 = fixture.debugElement.query(By.css('h1')).nativeElement;            // preferred
 
-      let h1 = fixture.debugElement.query(el => el.name === 'h1').nativeElement;  // it works
-
-          h1 = fixture.debugElement.query(By.css('h1')).nativeElement;            // preferred
-
-      expect(h1.innerText).toMatch(/angular 2 app/i, '<h1> should say something about "Angular 2 App"');
-    });
-
-  })));
+    expect(h1.innerText).toMatch(/angular 2 app/i, '<h1> should say something about "Angular 2 App"');
+  });
 });